+++ pyincore_incubator/analyses/cohortbuildingdamage/__init__.py
@@ -1,0 +1,1 @@
+from pyincore_incubator.analyses.cohortbuildingdamage.cohortbuildingdamage import CohortBuildingDamage